scrapy/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

psycopg2-binary==2.9.9
python-dotenv==1.0.1
Scrapy==2.11.2
